Add new decaf377-min crate
The problem with trying to migrate the old crate to no-std is that the
public API directly uses decaf377 stuff, so there will be
incompatability with the workspace as long as that isn't migrated.

Soon (TM), we will migrate the workspace over to the latest version of
decaf377, which will then allow this crate to replace the other one.